backend/inference/vllm_handler.py
"""
vLLM Inference Handler

Handles inference using vLLM framework for GPU-accelerated models.
"""

from typing import List, Dict, Optional, AsyncIterator
import logging
from dataclasses import dataclass

logger = logging.getLogger(__name__)


@dataclass
class InferenceRequest:
    """Request for model inference."""
    messages: List[Dict[str, str]]
    temperature: float = 0.7
    max_tokens: int = 4096
    top_p: float = 0.9
    stop: Optional[List[str]] = None
    stream: bool = False


@dataclass
class InferenceResponse:
    """Response from model inference."""
    content: str
    model: str
    usage: Dict[str, int]
    latency_ms: float
    finish_reason: str = "stop"


class VLLMHandler:
    """
    vLLM inference handler for large models.

    Supports:
    - Continuous batching
    - PagedAttention for memory efficiency
    - Streaming responses
    - Multi-GPU inference
    """

    def __init__(
        self,
        model_name: str,
        tensor_parallel_size: int = 1,
        gpu_memory_utilization: float = 0.90,
        max_model_len: int = 8192
    ):
        """
        Initialize vLLM handler.

        Args:
            model_name: HuggingFace model name
            tensor_parallel_size: Number of GPUs for tensor parallelism
            gpu_memory_utilization: GPU memory utilization fraction
            max_model_len: Maximum model context length
        """
        self.model_name = model_name
        self.tensor_parallel_size = tensor_parallel_size
        self.gpu_memory_utilization = gpu_memory_utilization
        self.max_model_len = max_model_len
        self._engine = None

    def load(self):
        """
        Load model using vLLM.

        This is a placeholder - actual implementation would use:
        from vllm import LLM, SamplingParams
        """
        logger.info(f"Loading vLLM model: {self.model_name}")

        # Placeholder for actual vLLM initialization
        # self._engine = LLM(
        #     model=self.model_name,
        #     tensor_parallel_size=self.tensor_parallel_size,
        #     gpu_memory_utilization=self.gpu_memory_utilization,
        #     max_model_len=self.max_model_len,
        #     trust_remote_code=True,
        # )

        logger.info(f"vLLM model loaded: {self.model_name}")

    async def generate(self, request: InferenceRequest) -> InferenceResponse:
        """
        Generate response using vLLM.

        Args:
            request: Inference request

        Returns:
            Inference response
        """
        import time
        start_time = time.time()

        # Convert messages to prompt
        prompt = self._format_messages(request.messages)

        # Placeholder for actual vLLM inference
        # sampling_params = SamplingParams(
        #     temperature=request.temperature,
        #     max_tokens=request.max_tokens,
        #     top_p=request.top_p,
        #     stop=request.stop,
        # )
        #
        # outputs = self._engine.generate([prompt], sampling_params)
        # generated_text = outputs[0].outputs[0].text

        # Placeholder response
        generated_text = f"[vLLM Response from {self.model_name}]"

        latency_ms = (time.time() - start_time) * 1000

        return InferenceResponse(
            content=generated_text,
            model=self.model_name,
            usage={
                "prompt_tokens": len(prompt.split()) * 1.3,  # Rough estimate
                "completion_tokens": len(generated_text.split()) * 1.3,
                "total_tokens": len((prompt + generated_text).split()) * 1.3,
            },
            latency_ms=latency_ms
        )

    async def generate_stream(
        self,
        request: InferenceRequest
    ) -> AsyncIterator[str]:
        """
        Generate streaming response.

        Args:
            request: Inference request

        Yields:
            Response chunks
        """
        # Placeholder for streaming implementation
        # In actual implementation, would use vLLM's async streaming API

        response = await self.generate(request)
        yield response.content

    def _format_messages(self, messages: List[Dict[str, str]]) -> str:
        """
        Format messages into a prompt string.

        Args:
            messages: List of message dicts with 'role' and 'content'

        Returns:
            Formatted prompt
        """
        prompt_parts = []

        for message in messages:
            role = message.get("role", "user")
            content = message.get("content", "")

            if role == "system":
                prompt_parts.append(f"System: {content}")
            elif role == "user":
                prompt_parts.append(f"User: {content}")
            elif role == "assistant":
                prompt_parts.append(f"Assistant: {content}")

        prompt_parts.append("Assistant:")
        return "\n\n".join(prompt_parts)

    def unload(self):
        """Unload model and free resources."""
        logger.info(f"Unloading vLLM model: {self.model_name}")
        self._engine = None
