backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI/ML Frameworks
torch==2.1.2
transformers==4.37.0
sentence-transformers==2.3.1
numpy==1.26.3

# vLLM for GPU inference (optional, install separately)
# vllm==0.2.7

# llama.cpp Python bindings
# llama-cpp-python==0.2.32

# Unsloth for QLoRA fine-tuning (optional)
# unsloth==2024.1

# Redis and caching
redis==5.0.1
# redisvl==0.1.0  # Redis vector library

# Database
# psycopg2-binary==2.9.9  # PostgreSQL
# sqlalchemy==2.0.25

# HTTP and async
httpx==0.26.0
aiohttp==3.9.1
websockets==12.0

# Data processing
pyyaml==6.0.1
python-dotenv==1.0.0

# Monitoring and logging
psutil==5.9.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1

# Development tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
